llm/requirements.txt
aio-pika==9.5.7
aiofiles==25.1.0
aiormq==6.9.2
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Deprecated==1.2.18
distro==1.9.0
fast-depends==3.0.3
faststream==0.6.2
get-annotations==0.1.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-tools==1.71.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.11.1
multidict==6.7.0
openai==2.6.1
packaging==25.0
pamqp==3.3.0
propcache==0.4.1
protobuf==5.29.5
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
yarl==1.22.0
